Update dependencies (#98)
* feat: support SEP-07 callbacks

using existing multisig code as much as possible, so
when wallet recieves SEP-07 requests with callback it stores a new signatureRequest
to activate multisig flow (even if there is just one signature).

When callback is provided the wallet won't publish the tx to the network (according to SEP07 spec).

Different cases handled by different parts of the code (Solar's design):
1. opening web+stellar:tx link by click or QR code scan
2. opening web+stellar:pay link by click or QR code scan
3. opening web+stellar:pay link from Payment Dialog using camera scanner

* executed `npm audit fix`

* remove babel. ES modules can be imported by node since 2020

* run npm audit fix on cordova project

* deps: bump dotenv-cli to latest major 2.0.0 → 8.0.0

I don't see dotenv is ever userful in the projects (no env vars in cordova/ ?)

* deps: remove isomorphic-fetch

In 2025 fetch is already supported by all the major browsers

* deps: bump jsonwebtoken 8 → 9

No migrations needed as we use only JWT.decode for SEP-10 code
https://github.com/auth0/node-jsonwebtoken/wiki/Migration-Notes:-v8-to-v9

* deps: upgrade electron 19 → 34 😱

Went through all the release notes, possible problems could be with:
- Macos: window drag (-webkit-app-region: drag, see: https://www.electronjs.org/docs/latest/breaking-changes#behavior-changed-draggable-regions-on-macos)
- Windows: context menu (I had no idea Solar/Sunce has context menu)
- Windows: menu bar will be hidden during fullscreen on Windows

Ref: https://www.electronjs.org/docs/latest/breaking-changes

Worth testing:
- notarize feature for Apple builds (electron-notarize)

* deps: using @suncewallet/stellar-transfer instead of @satoshipay/stellar-transfer

The former is a fork of the latter with updated dependencies

* deps: using @suncewallet/stellar-sep-10 instead of @satoshipay/stellar-sep-10